requirements.txt
tensorflow
streamlit
numpy
Pillow
matplotlib
scikit-learn